refactor: #63 정보 입력 공간 분리
