+++ src/components/index.ts
@@ -1,7 +1,7 @@
 import Header from './header/Header';
 import Subtitle from './title/subtitle/Subtitle';
 import Dot from './dot/Dot';
-import InputDropdown from './inputDropdown/InputDropdown';
+import InputDropdown from './inputDropdown/inputDropdown/InputDropdown';
 import CustomSelect from './customSelect/CustomSelect';
 import MeeteamTag from './tag/MeeteamTag';
 import MemberSelect from './memberSelect/MemberSelect';
@@ -79,6 +79,7 @@ import ManageRecruitCard from './recruit/ManageRecruitCard';
 import Status from './meeteam/status/Status';
 import Filter from './meeteam/filter/Filter';
 import Dropdown from './dropdown/Dropdown';
+import DateSelect from './dateSelect/DateSelect';
 import DeadlineSelect from './dateSelect/DeadlineSelect';
 import Icon from './meeteam/icon/Icon';
 import ApplyInfomation from '../pages/recruit/RecruitDetailPage/steps/ApplyInfomation';
@@ -90,6 +91,7 @@ import Pagination from './pagination/Pagination';
 import NaverLogin from './naver/NaverLogin';
 import Create from './header/Create';
 import TitleAndIntro from './title/titleAndIntro/TitleAndIntro';
+import InputContainer from './inputDropdown/inputContainer/InputContainer';
 
 export {
 	Header,
@@ -175,4 +177,6 @@ export {
 	Pagination,
 	Create,
 	TitleAndIntro,
+	DateSelect,
+	InputContainer,
 };
